boss hp ui , tightening circle anim

+++ Memorize/Boss.h
@@ -40,7 +40,7 @@ public:
 	virtual void OnTakeDamage(float damageAmount, struct DamageEvent const& damageEvent, class Controller* eventInstigator, Actor* damageCauser) override;
 
 	//Signal for UI
-	Signal<int>* OnHPChanged;
+	Signal<float>* OnHPChanged;
 
 };
